requirements.txt
django-autoslug<1.9.7
